chore(bld): move setOsxConfig function out
